+++ tests/canvas_frontend_live_feed_test.js
@@ -45,6 +45,7 @@ function buildHarness(allCards, filteredCards) {
     ${helpers}
     ${extractFunction("markAsNew")}
     ${extractFunction("syncChildren")}
+    ${extractFunction("renderTemplateOrBlocks")}
     ${extractFunction("populateLiveFeedItem")}
     ${extractFunction("renderLiveFeed")}
     ${extractFunction("updateCardElement")}